requirements.txt
astroid==2.3.2
autopep8==1.4.4
click==7.1.2
Flask==1.1.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
paho-mqtt==1.5.1
pycodestyle==2.5.0
pylint==2.4.3
six==1.12.0
typed-ast==1.4.0
virtualenv==16.7.8
Werkzeug==1.0.1
wrapt==1.11.2
